requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Data analysis and manipulation
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Financial data and analysis
yfinance==0.2.28
ta==0.10.2
pandas-ta==0.3.14b0
alpha-vantage==2.3.1

# Machine learning
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0

# Technical analysis
# talib-binary==0.4.26  # Removed - package doesn't exist
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Web scraping and sentiment
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
nltk==3.8.1

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

